+++ enterprise/internal/database/mocks_temp.go
@@ -6880,6 +6880,9 @@ type MockEnterpriseDB struct {
 	// OrgsFunc is an instance of a mock function object controlling the
 	// behavior of the method Orgs.
 	OrgsFunc *EnterpriseDBOrgsFunc
+	// PermissionsFunc is an instance of a mock function object controlling
+	// the behavior of the method Permissions.
+	PermissionsFunc *EnterpriseDBPermissionsFunc
 	// PermsFunc is an instance of a mock function object controlling the
 	// behavior of the method Perms.
 	PermsFunc *EnterpriseDBPermsFunc
@@ -7062,6 +7065,11 @@ func NewMockEnterpriseDB() *MockEnterpriseDB {
 		},
 		OrgsFunc: &EnterpriseDBOrgsFunc{
 			defaultHook: func() (r0 database.OrgStore) {
+				return
+			},
+		},
+		PermissionsFunc: &EnterpriseDBPermissionsFunc{
+			defaultHook: func() (r0 database.PermissionStore) {
 				return
 			},
 		},
@@ -7297,6 +7305,11 @@ func NewStrictMockEnterpriseDB() *MockEnterpriseDB {
 				panic("unexpected invocation of MockEnterpriseDB.Orgs")
 			},
 		},
+		PermissionsFunc: &EnterpriseDBPermissionsFunc{
+			defaultHook: func() database.PermissionStore {
+				panic("unexpected invocation of MockEnterpriseDB.Permissions")
+			},
+		},
 		PermsFunc: &EnterpriseDBPermsFunc{
 			defaultHook: func() PermsStore {
 				panic("unexpected invocation of MockEnterpriseDB.Perms")
@@ -7485,6 +7498,9 @@ func NewMockEnterpriseDBFrom(i EnterpriseDB) *MockEnterpriseDB {
 		},
 		OrgsFunc: &EnterpriseDBOrgsFunc{
 			defaultHook: i.Orgs,
+		},
+		PermissionsFunc: &EnterpriseDBPermissionsFunc{
+			defaultHook: i.Permissions,
 		},
 		PermsFunc: &EnterpriseDBPermsFunc{
 			defaultHook: i.Perms,
@@ -9766,6 +9782,105 @@ func (c EnterpriseDBOrgsFuncCall) Args() []interface{} {
 // Results returns an interface slice containing the results of this
 // invocation.
 func (c EnterpriseDBOrgsFuncCall) Results() []interface{} {
+	return []interface{}{c.Result0}
+}
+
+// EnterpriseDBPermissionsFunc describes the behavior when the Permissions
+// method of the parent MockEnterpriseDB instance is invoked.
+type EnterpriseDBPermissionsFunc struct {
+	defaultHook func() database.PermissionStore
+	hooks       []func() database.PermissionStore
+	history     []EnterpriseDBPermissionsFuncCall
+	mutex       sync.Mutex
+}
+
+// Permissions delegates to the next hook function in the queue and stores
+// the parameter and result values of this invocation.
+func (m *MockEnterpriseDB) Permissions() database.PermissionStore {
+	r0 := m.PermissionsFunc.nextHook()()
+	m.PermissionsFunc.appendCall(EnterpriseDBPermissionsFuncCall{r0})
+	return r0
+}
+
+// SetDefaultHook sets function that is called when the Permissions method
+// of the parent MockEnterpriseDB instance is invoked and the hook queue is
+// empty.
+func (f *EnterpriseDBPermissionsFunc) SetDefaultHook(hook func() database.PermissionStore) {
+	f.defaultHook = hook
+}
+
+// PushHook adds a function to the end of hook queue. Each invocation of the
+// Permissions method of the parent MockEnterpriseDB instance invokes the
+// hook at the front of the queue and discards it. After the queue is empty,
+// the default hook function is invoked for any future action.
+func (f *EnterpriseDBPermissionsFunc) PushHook(hook func() database.PermissionStore) {
+	f.mutex.Lock()
+	f.hooks = append(f.hooks, hook)
+	f.mutex.Unlock()
+}
+
+// SetDefaultReturn calls SetDefaultHook with a function that returns the
+// given values.
+func (f *EnterpriseDBPermissionsFunc) SetDefaultReturn(r0 database.PermissionStore) {
+	f.SetDefaultHook(func() database.PermissionStore {
+		return r0
+	})
+}
+
+// PushReturn calls PushHook with a function that returns the given values.
+func (f *EnterpriseDBPermissionsFunc) PushReturn(r0 database.PermissionStore) {
+	f.PushHook(func() database.PermissionStore {
+		return r0
+	})
+}
+
+func (f *EnterpriseDBPermissionsFunc) nextHook() func() database.PermissionStore {
+	f.mutex.Lock()
+	defer f.mutex.Unlock()
+
+	if len(f.hooks) == 0 {
+		return f.defaultHook
+	}
+
+	hook := f.hooks[0]
+	f.hooks = f.hooks[1:]
+	return hook
+}
+
+func (f *EnterpriseDBPermissionsFunc) appendCall(r0 EnterpriseDBPermissionsFuncCall) {
+	f.mutex.Lock()
+	f.history = append(f.history, r0)
+	f.mutex.Unlock()
+}
+
+// History returns a sequence of EnterpriseDBPermissionsFuncCall objects
+// describing the invocations of this function.
+func (f *EnterpriseDBPermissionsFunc) History() []EnterpriseDBPermissionsFuncCall {
+	f.mutex.Lock()
+	history := make([]EnterpriseDBPermissionsFuncCall, len(f.history))
+	copy(history, f.history)
+	f.mutex.Unlock()
+
+	return history
+}
+
+// EnterpriseDBPermissionsFuncCall is an object that describes an invocation
+// of method Permissions on an instance of MockEnterpriseDB.
+type EnterpriseDBPermissionsFuncCall struct {
+	// Result0 is the value of the 1st result returned from this method
+	// invocation.
+	Result0 database.PermissionStore
+}
+
+// Args returns an interface slice containing the arguments of this
+// invocation.
+func (c EnterpriseDBPermissionsFuncCall) Args() []interface{} {
+	return []interface{}{}
+}
+
+// Results returns an interface slice containing the results of this
+// invocation.
+func (c EnterpriseDBPermissionsFuncCall) Results() []interface{} {
 	return []interface{}{c.Result0}
 }